[FIX] project_enterprise: copy pill in task gantt view
Before this commit, since the recent changes to allow the duplication in
batch, the duplication of a pill in task gantt view no longer works and
raised a traceback. The reason is because the copy will now return an
array and so the `openPlanDialogCallback` function will do a for each on
that array to display eventual warnings, that override is in fact only
when smart task scheduling is called and not for the other actions
(copy, reschedule, etc).

This commit checks if the result given in parameter for `openPlanDialogCallback`
is not falsy and not an array before doing a for each on the object to
display the eventual warnings to display.

closes odoo/enterprise#57521

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>