[FIX] voip: user is stuck in busy state after a missed call
When an incoming call is cancelled by the remote party, the current
session is not properly cleared, causing Odoo to reject all subsequent
calls, assuming the user is busy, since the previous session is still
present.

This commit fixes the problem by properly clearing the current session
when a CANCEL message is received.

closes odoo/enterprise#57525

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>