[FIX] account_intrastat: Remove warnings in group by
This commit(https://github.com/odoo/enterprise/commit/26fc172004d880f5e941c78cbdc18a29d2db5134) introduced
a bug. This bug is caused when several lines that should be
grouped in one line have different warnings. Before this fix, it
generates several grouped lines instead of one.
This is causing a javascript issue where the generic unique line id
is duplicated (and it shouln't be the case).

To fix it, this commit removes the group by in the _build_query_group
function and accepts null value in the ARRAY_AGG. When the code is checking
if we have warnings, we remove the null values to only keep ids.

opw-3751773

closes odoo/enterprise#57540

X-original-commit: f573074fdcdada52545f5730f5bcded607ddd550
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>