[FIX] sale_subscription: recompute some optnl SOLs
**Current behavior:**
Given a portal preview of a recurring sale order with an order
line (for a recurring product) and a distinct optional product
which is also recurring, if the optional product is added to
the cart and then the order is brought back into the editor
view, should the recurring plan for the order be changed, the
optional product (which is now added to the order) will not
have its price adjusted accordingly.

**Expected behavior:**
The optional product that has been added to the order will have
its price adjusted to match the recurring plan selected for the
sale order.

**Steps to reproduce:**
1. Create two distinct recurring products and assign them
     subscription prices for the same recurring plan timetable

2. Create a new sale order, adding one product to the order and
     the other to the optional products tab

3. Preview the sale order and add the optional product to the
     cart, then return to the editor view

4. Select a new recurring plan for the sale order where each
     product has a different set price than the one which is
     currently displayed on the order lines

5. The product which was initially added in an order line will
     have its price adjusted accordingly, while the product
     which was added from the optional products section will
     remain static.

**Cause of the issue:**
In sale_management, the _compute_price_unit() method is
extended such that it filters SOLs which have order options
associated with them.

**Fix:**
Override the filter method from sale_management so that lines
which are associated with recurring products will end up having
their prices updated.

opw-3698798

closes odoo/enterprise#55710

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>