[FIX] timesheet_grid: fetch last validated timesheet date when data
Before this commit, since 8df2e440 to avoid fetching the last validated
timesheet date when the sample data is displayed, when the user changes
the range of the grid view or remove a default filter and get some data.
A traceback is occured because `lastValidationDatePerEmployee` is
undefined. That attribute in timesheet grid renderer is indeed undefined
since we did not fetch the data to populate that attribute.

This commit makes sure to fetch the last validated timesheet data once
the user leaves the sample data.

closes odoo/enterprise#57535

X-original-commit: 8df2e440
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>