[FIX] hr_appraisal: add attachment_ids arg in actn
**Current behavior:**
Modifying an email template associated with the Appraisals app
will not result in that attachment being sent when a message is
composed using that the template.

**Expected behavior:**
The attachment(s) of a certain template will be sent when an
email is triggered that references that template.

**Steps to reproduce:**
1. Modify an HR Appraisals email template so that it has some
     file attachment

2. In the Appraisals app, create a new appraisal that will use
     the modified email template

3. Send the email, observe in the chatter sidebar or in the
     list of emails in settings that the attachment linked to
     the relevant template was not sent

**Cause of the issue:**
The attachment_ids field of the request.appraisal model was not
passed as a kwarg to the action_invite() method, which is where
the email gets composed and posted.

**Fix:**
Add the missing attachment_ids argument to this method.

opw-3644820

closes odoo/enterprise#57681

X-original-commit: 3fde6464749bd4261ba70a4460cc02049cf5de5a
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>