[FIX] sale_subscription: remove upsells information in the report
Steps to reproduce:
-------------------
- create a subscription with a recurring product (which costs 100);
- confirm it and create an invoice;
- create an upsell for the subscription;
- add the same product with a discount of 30%;
- go to Sales report and search the product (product variant);

Issue:
------
The report displays 300 for "Untaxed Total" instead 170.

Cause:
------
The report takes into account the upsells and when they are validated,
the information is put back on the original subscription.
As a result, the information is duplicated.

Solution:
---------
Do not include upsells in the sales report.

Note:
-----
With the fix, the value is 200 and not 170, because it is not possible
to take upsell discounts into account in the sales report
(we have to use the subscriptions report).

opw-3646388

closes odoo/enterprise#58369

X-original-commit: de07949a6c52529c0bd00edc8c76ec3836be3547
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>