[FIX] web_studio: approvals with existing entry
Have an approval flow rather complex in four steps:
- rule 0: notification order 1, exclusive
- rule 1: notification order 2, exclusive
- rule 2: notification order 2, exclusive
- rule 3: notification order 3, exclusive, has a responsible

Obviously, each validation should be done by a different user (and not an admin), because all rules are exclusive.
Also, each rule has a responsible in the example below, but really only rule 3's is part of the real
use case and is the object of the fix.

Validate 0: 1 and 2 receives an activity
Validate 1: activity of 1 is removed
validate 2: activity of 2 is removed, 3 receives an activity (<=====)

Before this commit, 3 did not receive an activity. This was because, in this flow
the rule 3 is evaluated for validation because every rule before it had been validated.

However, in that part of the code, we ensure that no rule below it in notification order is not validated, but
we do it in the context of the current user.

So, when validating rule 3, we couldn't see the validations for at least of previous rule that another part of the code
had seen before.

After this commit, the rule 3 receives an activity, and the whole flow works as expected.

opw-3752133

closes odoo/enterprise#57604

Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>