[FIX] helpdesk: provide fallback default val
**Current behavior:**
When creating a helpdesk team, if you check Automatic
Assignment, delete the default member, then uncheck Automatic
Assignment and try to save, you will be prompted to input at
least one member in the team depsite the input UI disappearing.

**Expected behavior:**
The form should not require a value for a field when the UI
does not permit one to be entered.

**Steps to reproduce:**
1. In the Helpdesk app, go to configuration -> Helpdesk Teams

2. Create a new team, select Automatic Assignment, delete the
     member, then uncheck Automatic Assignment

3. Try to save the record

**Cause of the issue:**
member_ids is a required attribute of the helpdesk.team model
despite the form view hiding the selection box when Automatic
Assignment is unchecked (implying it is not required).

**Fix:**
Create an _onchange() method which populates the member_ids
field with a default value when the user checks/unchecks
Automatic Assignment.

opw-3672587

closes odoo/enterprise#57653

X-original-commit: 9634f9dfc698399a617cc3e0a7a24a5c0bbeec1a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>