[FIX] voip: call stuck in invitation view
Description:
============
If the execution of the code is abruptly interrupted during an incoming
call invitation (e.g. by closing the page), when the call is reopened in
the softphone, the correspondence details of the call get stuck in the
invitation view.

This commit prevents this from happening by also checking for the
presence of an actual session as a condition to display the call
invitation view.

How to reproduce:
=================
1. Configure a database with VoIP in production mode
2. Receive an incoming call invitation and refresh the page without
   accepting/rejecting it
3. Re-open the call and notice that it's stuck on the invitation view
   :((

Task-3770997

closes odoo/enterprise#57618

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>