[FIX] account_asset_fleet : store the correct cost value in fleet service
Steps to reproduce:
	1- Install Accounting, Fleet modules
	2- Create a bill in accounting with a different currency than the company's default, and add a line with a chosen vehicle_id.
	3- Go to the chosen vehicle in Fleet module
	4- Navigate to the service created for this bill

Current behavior before PR:
If we create a bill for a vehicle using a different currency than
the company's default. The fleet service that will be created
will be having the company's currency but the value will be the amount
in the currency used in the bill

Desired behavior after PR is merged:
We now create the fleet service using the value in debit
not the unit price or the price subtotal.

opw-3734743

closes odoo/enterprise#61408

X-original-commit: 2a73cd04ecaa7db3aeb84821499d806bcac81496
Related: odoo/odoo#163228
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>