[FIX] account_report: shring the aged balance reports for pdf export
Steps to reproduce:

- Create a vendor bill for a partner with a super long name.
- Go to Accounting > Reporting > Partner reports > Aged Payable
- Generate the pdf

> If the partner's name is long enough the last columns will not appear

Cause of the issue:

Pdf are generated in Odoo using wkhtmltopdf. This one takes an html file
renders it in Qtweb (shrinking its content to fit a single screen) and
making a screenshot to generate a pdf. However, there is a limit to the
shrinking applied to the html body. Therefore, if the content is too
wide part of it will not appear on the pdf (this can easily been
observed using Qtweb (but not with chrome)). Between 16.4 and 17.0
there was a refactoring of the way account reports are handled in Odoo
see commit 2d77434 and commit b8209e2. Because of this refactoring,
there is currently no width limit for cells of the table used in the
body of account reports. As such, if the content of the first column
is too wide, every other column will be pushed out of the screen before
the pdf is generated. This is exactly what happens here.

Fix:

We added a max width to cells and a wordbreak option on the css file
used for the html in the pdf export of every account reports. This
ensures that the content of any cell will not push the other cells
outside of the screen before the screen shot.

Second Issue:

Even with this issue out of the way, it is still possible for the
content of the aged partner balance reports to be too wide after the
shrink of the hmtl if you use a currency whose symbol is too long.
The reason is the same as before coupled with the fact that these
reports contain too many columns displaying the currency symbol of the
company.

Second part of the fix:

To fix this second issue, we use a customm css class for the export of
the two problematic reports (aged payable and aged receivable) when the
currency symbol is too wide.

opw-3749625

closes odoo/enterprise#60443

X-original-commit: 2b7c0ba58b7f2923e35a7f1c9bf1a212934022ce
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>