[FIX] mrp_workorder: component_uom_id in compute
In the `quality.check` model, before this commit, the field
`component_uom_id` was updated in the `_compute_component_data` compute
method. The issue is: this field is a related field and should not be
updated in a compute.
That was causing some issue while reading the `component_remaining_qty`
field, when opening Shop Floor for example.

X-original-commit: 8aa89f219ac566c0742f7668973b6f5e99af671f
Part-of: odoo/enterprise#58074