[FIX] sale_subscription_stock: copy qty_delivered from upsell to parent
In sale subscription, if you upsell a recurring consumable product, it will
create a delivery order on the upsell. After this delivery order is confirmed,
the upsell will have its qty_delivered set. In order to keep the parent sub up
to date, we should also set the qty_delivered of the upsell line on the parent
sub line.

task-3689908

closes odoo/enterprise#55613

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>