[FIX] l10n_mx_edi: import: fix partner creation and detection logic
There are currently the following problems when importing an invoice
with an unknown partner. A new partner is created in this case.
For invoices with an unknown domestic partner (no special VAT values):
- (1) There may be a VAT validation error.
- (2) The "CFDI to public" field (l10n_mx_edi_cfdi_to_public) on the
      invoice will be set to True. It should be False.
- (3) Invoices created using a created partner as customer
      are missing the (required) DomicilioFiscalReceptor value in
      the Receptor section (zip code).
For invoices with an unknown foreign partner (no special VAT values):
- (4) The partner is created w/o country or VAT value (since only
      the special VAT / Rfc value XEXX010101000 is used).
      Thus the detection logic applied to the partner to decide
      whether we are in the domestic non-public, domestic public or
      foreign case does not work correctly.

The issues are solved after this commit by the following changes
- (A) The country is set to Mexico for all created partners that are not
      from a foreign invoice (XEXX010101000).
- (B) The detection logic to decide the case is updated.
      Also see (C).
- (C) The "Foreign Customers" fiscal position is set for partners
      created during the import of foreign invoices.
- (D) The zip code is set for invoice customers.
      It is parsed from the (required) DomicilioFiscalReceptor value
      in the Receptor section.

Note on (B) / (C):
We still want to assume that partners without a country are Mexican
partners. This is done since (many) Mexican users will not set an
explicit country for Mexican customers.
(One would e.g. not put the country for domestic letters either.)
This is why we use the "Foreign Customers" fiscal position (C).

### Details

(1)
Due to the missing country the wrong VAT validation is used.
The same error can be reproduced on a runbot when creating a new
customer without country and VAT DEA040805DZ4
(works after setting the country to Mexico).

(2) / (4)
Due to the missing country on the partner the invoice is wrongly
interpreted as a domestic invoice to public (XAXX010101000).
This leads to the wrong "CFDI to public" value.

### task

task-3731058

closes odoo/enterprise#58776

X-original-commit: f1936e759e8e5f7dc267a342814fe1e836d23704
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>