[FIX] industry_fsm_report: use `in` operator when right leaf is an array
Before this commit, a warning is logged in the server logs because `=`
operator is used with an array as right leaf in a domain used in a
search in `_compute_worksheet_template_id` defined in `project.project`
model. The operator used should be `in` instead of `=`, that's why we
got that warning.

This commit fixes the issue by changing the operator used inside that
domain to use `in` instead of `=`.

runbot-58374

closes odoo/enterprise#57725

X-original-commit: 2b12d5e8
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>