[FIX] account_reports: fix traceback in Partner Ledger
When activating the groupby_prefix parameter for the Partner Ledger, a
Traceback can happen if clicking on Unfold All and some of the partners
are archived. That's because the search on res.partner in
_custom_unfold_all_batch_data_generator injects the active ir.rule
whereas the search in _query_partners does not. To fix that,
pass active_test=False to the context.

To reproduce:

 - Install account_reports and contacts with demo data
 - Archive Azure Interior
 - Set the parameter account_reports.partner_ledger.
   groupby_prefix_groups_threshold to 2
 - Go to Partner Ledger
 - Click on Unfold All
 - A Traceback is raised

Also changes the search domain operator from ilike to =ilike.

opw-3703069

closes odoo/enterprise#60983

X-original-commit: a55a568495184494453959a82f75351d8dc8b615
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>