[FIX] helpdesk: provide fallback default val
**Current behavior:**
When creating a new helpdesk team, if you have selected Random
or Balanced as an Assignment Method and delete the default
user in the avatar selection box then switch to the Manual
assignment method option, you will simultaneously be unable to
assign a user (as the selection box disappears) and required to
provide at least one user before being saving.

**Expected behavior:**
The form should not require a value for a field when the UI
does not permit one to be entered.

**Steps to reproduce:**
1. In the Helpdesk app, go to configuration -> Helpdesk Teams

2. Create a new team, select either Random or Balanced as the
     Assignment Method, delete the default value, then switch
     back to Manual assignment

3. Try to save the record

**Cause of the issue:**
member_ids is a required attribute of the helpdesk.team model
despite the form view hiding the selection box when the Manual
assignment option is selected (implying it is not required).

**Fix:**
Create an _onchange() method which populates the member_ids
field with a default value when the user switches to Manual
assignment with no member_ids assigned.

opw-3672587

closes odoo/enterprise#57778

X-original-commit: a88e8835581778420c18a99f62e102046cc345c8
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>