[IMP] mail, *: chatter without activity ids in view arch
Before this commit:
Detecting if a chatter has activities or not, was doing by checking the
availability of `activity_ids` as a child of form compiler node.

After this commit:
Every model that inherits the `mail.activity.mixin` is an instance of this
registry and put `activity_ids` as a child of `oe_chatter` in a form view
without inheriting from this mixin is actually useless. So, to check if a thread
 in a chatter has `activities` or not, we can simply check the related model

Part-of: odoo/enterprise#57020