[IMP] documents: hide/disable forbidden actions
We disable the "upload" button when the user doesn't have write access to the
selected workspace.

We ensure that the workspace field in the inspector does not allow to select
non-writable workspaces, as the user cannot move selected items to those
workspaces.

We extend the feature developed in odoo/enterprise#38430 to the trash and
non-writable workspaces, to avoid having an error when dropping files in the
trash or a non-writable workspace.

Task-3707728

closes odoo/enterprise#57917

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>