[FIX] mrp_workorder: manual consumption if check
This commit marks a move as manual consumption in case one of its move
line is linked to a quality check.
Since a "register consumed materials" check asks to manually enter the
quantity, it makes sense linked move/move line is marked as manual as
well, and before this commit, as it was not the case, thus there is some
issue about the quality check quatity.
The QC qty was set to 0 because the linked move line was already picked
(picked at the workorder start since the move was not on manual
consumption) and so, the way we compute the QC qty, we substract picked
move line qty.  # Please enter the commit message for your changes.
Lines starting.

X-original-commit: 8f777e1b94b92720328265cbc53c7923f68a16b4
Part-of: odoo/enterprise#58074