[IMP] sale_subscription: invite customers to automate future payments
With this improvement, during a subscription payment with payment methods which support tokenization it is now possible to invite customers to automate their future payments by saving the payment token to the subscription. This update will provide to the customer a better experience while handling payments.

Besides that, before this commit it was not possible for the customer removing a token linked with subscriptions which didn't have online required payment. Now, it is possible to remove them in case the customer change his mind.

closes odoo/enterprise#36656

Task-id: 3144077
Related: odoo/odoo#116202
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>