[IMP] whatsapp: improve variable and button constraints
In order to lessen the number of errors raised when
using and editing the whatsapp.template form view, and
to lighten its use, we remove two constraints linked to
the buttons.

First, we remove the constraint on the validation of
the call number on the buttons. In order to still inform
the user, we display a warning message as an alert at the
top of the form view, when such a button contains an ill
formated number. Computed fields are added to do so.
The number is first checked without setting country,
then if not formatted properly we check a second time,
this time using user's country, or company's country.
This avoids displaying the warning message for numbers
correctly formatted for that country, which may be one
of the base use cases.

Second, we remove the constraint on the formatting of url
in 'url' whatsapp buttons. Instead, we add an onchange method
on the url entered by the user. If it does not start with
http or https, we add the prefix 'https' automatically.
Also, the placeholder is restored as it was hanging in
the template without matching any opening bracket.

The test checking constraints is deleted and another is
added to assert the points above.

A new wording is used on the variable constraint.

Task-3653289

Part-of: odoo/enterprise#57526