[FIX] l10n_au_hr_payroll: remove match case
The pylint version currently used on runbot does not
support match like we used in the previous
commit, but it was still green and got merged.
In order to ensure consistency with the style checks,
we revert that change and fix the styling issues shown
by pylint.

task # 3639270

Part-of: odoo/enterprise#57884