[FIX] stock_barcode: split when scan expect dest
In a picking, when a destination is scanned, if the scanned line is not
complete, the line is splitted: the processed qty goes to the scanned
destination and the remaining qty still goes to the previous expected
destination. That said, if the scanned destination is the move line
expected location, the line is not split, which can be annoying in some
case.

How to reproduce:
- Active multi location;
- Create a receipt for 4x product A and confirm it;
- Open the receipt in the Barcode App;
- Scan 2x product A and then scan WH-STOCK as the destination
  -> Nothing happens.

If the picker wants to move 2 product A in WH/Stock and 2 product A in
Shelf 1, they have no easy way to do it if they scan WH-Stock first.
If they scan Shelf 1 first, then the product A line will be splitted and
they will get:
- 2/2 product A going to WH/Stock/Shelf 1;
- 0/2 product A going to WH/Stock.

But if they scan WH-Stock first, they will stay with only one line:
- 2/4 product A going to WH/Stock.

In such case, no other choice than create a new line through the "Add
Line" button (form view) and so, the reservation won't be split between
the two lines.

OPW-3774095

closes odoo/enterprise#58118

X-original-commit: ad18f9dc93a7c2d4053cd5d7e6cb0024e0fbd01b
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>