[FIX] project_enterprise: allow to extend taskGanttView
Before this commit, the taskGanttView object was not exported and so if
someone would like to override that object, he will have to get that
object in the views registry.

This commit allows to export taskGanttView from that file to easily use
it to create another Gantt JS based on that one instead of getting the
object from the registry.

task-3776786

closes odoo/enterprise#57947

X-original-commit: 6738c6c8cd60699bed6c46523f7d46701d87853d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>