[FIX] sale_external_tax: tour is failing number
The tour is failing rarely and undeterministically.
It seems like the reason is because 2 steps from the tour are done
before a request is done (and so the number of the call to the function).
We add a step to enforce that the request is made before the next
step of the tour.
We also enforce that we don't do the last action before closing the tour.

In ulterior versions, the tour can be stuck when adding a product.
Our guess is that the button is clicked before the page is fully loaded.
The tour would then consider it did click the button, while the action didn't happen.

We add a check to wait for the page to be fully loaded before clicking.

Linked to runbot error 61458, 63532, 59031 (and happened before already).

closes odoo/enterprise#63229

X-original-commit: fb13121e9ab43745eb90fb9e62bd4cec1bc869bb
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>