[FIX] delivery_starshipit: exact_price
The current code was made in response to an issue
where the exact_price value would not always be in
the response of /orders, which would cause issues
and inconsistencies.

As of now, this issue seems to not happen anymore
and we can thus revert to the original way of
handling that which uses less api calls and a faster
flow.

closes odoo/enterprise#58040

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>