[FIX] delivery_starshipit: exact_price
The current code was made in response to an issue
where the exact_price value would not always be in
the response of /orders, which would cause issues
and inconsistencies.

As of now, this issue seems to not happen anymore
and we can thus revert to the original way of
handling that which uses less api calls and a faster
flow.

closes odoo/enterprise#58072

X-original-commit: 564d9e1d5575dc327f7e0302641fa99a2a5a25ac
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>