[FIX] hr_payroll: fix issue for generating deleted declaration
Steps to reproduce:

- Create a payroll declaration sheet
- Populate the declaration
- Delete the declaration sheet
- Create another payroll declaration sheet
- Populate the declaration again
- Try to generate the declaration pdfs

Current behaviour:

- Error raised

Expected behaviour:

- Generate declaration pdfs successfully

Explanation:

When deleting declaration sheet, declarations will not be deleted.
They remained in the database. Therefore when excuting the generate
pdfs scheduled action, the system will search the declarations and
try to browse the deleted declaration sheet. As a result, error
raised.

X-original-commit: 2f514e4ef9b9b3f945b1ce80770c2f133e0ea142
Part-of: odoo/enterprise#58014