[FIX] appointment: restore slot drag
Purpose
=======
Restore the possibility to resize and drag and drop
the slots when using the creation mode in the calendar.

Specifications
==============
The calendar events edition (drag and drop/resize actions) were
previously limited to the participants of the event.
This prevented, for example, users from changing the time-off events
of other users. However this fix was applied on a method that is also
used by the calendar slot renderers. As slots don't have any
participants, their editable attribute was always set to false.

Fixing the issue by forcing the slots editable attribute to true.

Task-3756066

closes odoo/enterprise#58171

X-original-commit: b99a98e266610bea2eeb6932225b72e7e8f50d7a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>