[FIX] knowledge: redirect users to the article they were reading after signing in
When a user reads an article from the frontend view of Knowledge and
clicks on the "Sign in" button to sign in: internal users will be
redirected to the Odoo backend while portal users will be redirected
to the `/my` page.

The current redirection process is confusing because users may lose track
of the article they were reading. If people re-open Knowledge after being
redirected, they may no find the article they were reading as it could
be hidden in the sidebar.

When people click on the "Sign in" button from Knowledge, we assume that
they probably wanted to edit the article, change the permissions, add
new members, access their own workspace, etc. So, we will now redirect
the user to the article they were reading after they sign in.

With that change, portal users can still access the '/my' page from
Knowledge by clicking on the home icon of the sidebar and internal users
can view all apps of the backend by clicking on the home button of the
Odoo navbar.

task-3776350

closes odoo/enterprise#59161

X-original-commit: 55c6566ccd820145c07157cbd0f515b777e0b694
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>