[FIX] web_studio: fix test image crop undeterministic load error
Before this commit the test `test_image_crop` sometimes failed.
This was because the browser quit while the image of the image cropper was loading.
This operation caused an error, that was caught by the browser and then by the test infrastructure
but *after* the tour was marked as succesful.

After this commit, we wait a little bit longer, that is, until the cropper widget (which is a JQuery extension)
appears.

runbot-error-59063

closes odoo/enterprise#58024

X-original-commit: d6f0dbdb2922f1db7ceed256ff690e372c9c7948
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>