[FIX] industry_fsm_sale: don't invoice zero SOL from FS
Steps to reproduce:

- Go to Field Service app and create a new task.
- Use the smart button to access the catalog and add a product.
- Go back to the task and use the smart button to access the SO.
- Set the unit price of the added product to zero.
- Create the associated invoice.

> You are allowed to create the invoice and aline will be displayed for
the SOL corresponding to the products with a unit price of zero.

Notes: This is the expected behavior for invoices generated from regular
SO. However, AJU the PO of the Field Service application decided to
change this behavior for invoices generated from SO coming from field
services see 3266856 and commit 849241e (the SOL invoice status should
be 'no').

Cause of the Issue:

When you create an invoice from a sale order, the `invoiceable_lines`
are computed at this point:
https://github.com/odoo/odoo/blob/96a06df867fcef1fd24a1798bd5d6decc423933b/addons/sale/models/sale_order.py#L1251-L1258
by the `_get_invoiceable_lines` method. As one can see in this method:
https://github.com/odoo/odoo/blob/96a06df867fcef1fd24a1798bd5d6decc423933b/addons/sale/models/sale_order.py#L1220
if the `qty_to_invoice` of SOL is positive, the SOL will therefore
appear in the `invoiceable_lines` and the associated invoice line will
be created a few lines below to appear on the invoice:
https://github.com/odoo/odoo/blob/96a06df867fcef1fd24a1798bd5d6decc423933b/addons/sale/models/sale_order.py#L1276-L1279

Fix:

In order for the SOL to not generate an`invoiceable_line` we set the
`qty_to_invoice` to zero when the SOL comes from a field service task
and display a product with a unit price of zero.

Notes: To keep the flow coherent between the `qty_to_invoice` and the
`invoice_status` of SOL for field services SO and since `price_unit` is
a float that should not be compared using an `==` we also changed the
float comparaison of the `_compute_invoice_status` method
introduced in the commit 849241e.

opw-3776475

closes odoo/enterprise#58057

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>