[FIX] helpdesk_sale_timesheet: compute commercial partner for timesheet
Issue:
------
In the helpdesk application, when a timesheet is added to a helpdesk ticket,
it does not have a commercial partner
(we can add the field with studio to see this behaviour).
There may be one if a commercial partner exists for the helpdesk project.

This behaviour is contradictory to that found in project.

Solution:
---------
Add the `helpdesk_ticket_id` field invisibly in the view so that this field is in
the fields to be updated with the onchange.
It will be correctly updated during the onchange.
Override the compute method to obtain the `commercial_partner_id` field
in order to take the commercial partner from the helpdesk ticket
and not from the task.

opw-3759824

closes odoo/enterprise#58370

X-original-commit: 99de614811976c61d3779b00eb52f9e6712e2d28
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>