[FIX] knowledge: remove useless popover
This commit fixes an issue where the popover would open inside the comments
panel when replying to a comment with a small enough window, even though in
this situation it's not useful at all.

Now when the panel is open and we have a small window, the popover will not
open when replying to a comment.

task-3786282

X-original-commit: 421d05cc1b7e48b1228e9c8e79f0ece54583573e
Part-of: odoo/enterprise#61094