[FIX] knowledge: remove useless popover
This commit fixes an issue where the popover would open inside the comments
panel when replying to a comment with a small enough window, even though in
this situation it's not useful at all.

Now when the panel is open and we have a small window, the popover will not
open when replying to a comment.

task-3786282

Part-of: odoo/enterprise#58045