[FIX] hr_payroll: use env lang when emp lang false
**Current behavior:**
If a payslip report is printed for an employee with no set
language, it will be printed in English.

**Expected behavior:**
The primary fallback language should be that of the database
before going to the base level default of English.

**Steps to reproduce:**
1. Activate spanish and spanish MX languages on the DB, set
     all users' language to Spanish

2. Deactivate the default English language pack

3. In the Nómina (Payroll) app, go to `Recibos de nómina` ->
     `Todos los recibos de nómina`, then select any report and
     click the Imprimir (Print) button

4. Observe some sections are in English

**Cause of the issue:**
When an employee's language is not set, English is used as a
default.

**Fix:**
Look to use the environment's language if the Employee's is
False.

Ammended to fix other situations in hr_payroll where an employee
language is accessed without using `env.lang` as a primary
fallback.

opw-3747839

closes odoo/enterprise#58461

X-original-commit: a4937bbf454559bc68c38f037f84f0504c82e81a
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>