[FIX] account_accountant: adapting tests for branch changes
Description of the issue/behavior this commit fixes:

Due to changes done on this branch, some tests are now failing. It is not
due to an error but because the behavior has been modified. These tests
should succeed.

Desired behavior after this commit is merged:

The enterprise tests that are failing due to changes done in this branch
now succeed.

task-3642419

closes odoo/enterprise#61327

Related: odoo/odoo#163014
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>