[IMP] l10n_pe_edi: show spot details on invoice
In Peru we have the SPOT system, where a customer may have to pay a
withholding on the invoice he receives in a specific bank account.

Currently some of the required information is only available in the XML
of the electronic invoice, but not on the PDF representation.

In this commit, we add the withholding amount and the bank account
number on which they have to pay to the PDF, as well as format all the
information in a clear way.

task-3413744

closes odoo/enterprise#58049

Signed-off-by: Laurent Smet (las) <las@odoo.com>