[FIX] sale_planning: fix traceback when clicking on pill when no demo data in planning gantt view
Steps to reproduce:
-------------------
1. Install sale_planning (without demo data)
2. Create a service product with "Plan Services" activated
3. Create an SO with this product and confirm the SO
4. Click on "To Plan" stat button
5. Click on any pill to create a new slot
6. The traceback occurs

The traceback also occurs in the following scenario:
-------------------
1. Install sale_planning (without demo data)
2. Open Planning > Schedule > By Role
3. Click on any pill to create a new slot
4. The traceback occurs

Fix:
-------------------
When there is no demo data/records, the gantt view uses sample data.
In which for instance, the role_id field is randomly generated. So when clicking on
the pill to create a new slot with the given role_id, the corresponding
'planning.role' record is not found, resulting in a traceback.
To fix that, we check if the sample data is enabled, in which case we ignore
the "fake" ids given by the sample records.

task-3777485

closes odoo/enterprise#58420

X-original-commit: 81132c50c0926d3a6cf23dbea4dae09403023e79
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>