[FIX] knowledge: show panel's comments in small windows
This commit fixes an issue with the comments panel when using it with a small
window.

Before, when you opened the panel when in a small enough window, the comments
aren't properly displayed meaning that you only have empty boxes inside the
panel. This was caused by a styling issue inside of the template that applied
a wrong style to the comment boxes inside panel, which is not necessary inside
it.

Now, we modified the inline style in the template so that it is not applied
inside the panel, showing back those comments in all circumstances.

task-3786282

closes odoo/enterprise#58045

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>