[FIX] *: HOOT - unify white spaces in queryText
This commit replaces all white space representations in the `queryText`
and `queryAllText` helpers by a single "standard" white space character
(ASCII 32).

This is done to ease the comparison of strings in tests.

An optional `raw` option can still be passed to retrieve the original
text.

Some tours relying on queryText and queryAllText have been updated to
not rely on "nbsp" characters anymore.

closes odoo/enterprise#58135

Related: odoo/odoo#156255
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>