[FIX] mrp_workorder: employee hook cleaning
This commit was not asked in the task but some cleaning is required.

In employee_hook, the onBarcodeScanned function was not used anymore.
In addition, the get_employee_barcode python function is not related to a workcenter
so it has been made static.

Removed openEmployeeSelection, openRecord from employee_hook

Task-id: 3691130
Community: https://github.com/odoo/odoo/pull/154607
Part-of: odoo/enterprise#56927