[IMP] l10n_be_codabox: from_date optimization
Currently we send a list of ibans, and a single date from which to fetch the files.

This is a bit problematic if the user has multiple journals and one of which is not used.
In such a case, this old journal won't have transactions for months. Then we'll use the date
of this old journal as the one from which to start fetching. So a journal that is up to date
and gets documents everyday will still use this old date. Therefore we'll download lots of files
(3 months or historic) even though the vast majority will not be used (only skipped because already imported)

From now on, for each journal, we send the date of the last bank statement along with the IBAN
of the journal, so that on IAP's side, we can filter out older bank statements. This allows
faster fetching of the files.

task-id 3762743

closes odoo/enterprise#59051

X-original-commit: b9d95bb49588852f74fee285969c67b263a0b8a9
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>